Remove an extra level of indirection in memory context routines
The memory context object formerly had an option to replace the memory
routines with own implementation when used outside of BIND 9.  The
ability to actually replace these was removed some time ago as the
libraries are now considered internal to BIND 9, and the default set of
the routines was always used.  The extra layer of indirection has been
removed now as well.