WIP: Needs split
Part 1: Move the constructor/destructor to lib/isc/lib.c as there's a
bug in Apple Clang that prevents the priority in
__attribute__((constructor|destructor)) to work properly across the
modules, e.g. all constructors and destructors needs to be listed in a
single compilation unit.

Part 2: Add proper support for compiling with different memory
allocators, initially, we have these memory allocators available:

1. the current custom BIND 9 allocator, which is default
2. the system library memory allocator
3. jemalloc memory allocator
4. tcmalloc memory allocator from gperftools

The work isn't 100% complete, as the non-default memory allocators are
missing the accounting, so the memory limits don't work at the moment.

The other suboptimal part is allocating chunks for the mempool, basically
the mempool should be fixed separately, as allocating the memory as
individual chunks is very suboptimal, it's slow and it leads to memory
fragmentation.