tracing of active sockets and handles
If NETMGR_TRACE is defined, we now maintain a list of active
sockets in the netmgr object and a list of active handles in
each socket object; by walking the list and printing `backtrace`
in a debugger we can see where they were created. This will assist
in debugging reference counting errors.