fix LD_WRAP tests
I noticed that query_test and notify_test still had references to
isc_nmhandle_unref(), which hadn't been causing problems because they
weren't being run. this appears to be due to a logic reversal in
configure; the test program returns 0 if -Wl,--wrap is not supported.
after fixing this the unit tests needed to be cleaned up in order to
pass.