Run a minimal set of system tests on Windows for merge requests
Regular system test on Windows is run only for schedule-, web-, and
tag-triggered pipelines because the test suite on Windows is slow and
takes scarce resources. However, sometimes BIND is broken on Windows,
and we find out about that only after the code got merged. To identify a
potential breakage, a minimal subset of the system test(s) is run for
merge requests.