Add the reader-writer synchronization with modified C-RW-WP
This replaces the current custom reader-writer lock implementation based
on

  John M. Mellor-Crummey and Michael L. Scott.  1991.  Scalable
  reader-writer synchronization for shared-memory multiprocessors.  In
  Proceedings of the third ACM SIGPLAN symposium on Principles and
  practice of parallel programming (PPOPP '91). Association for
  Computing Machinery, New York, NY, USA, 106–113.
  DOI:https://doi.org/10.1145/109625.109637

(pseudocode available from:
  https://www.cs.rochester.edu/research/synchronization/pseudocode/rw.html)

with

  Irina Calciu, Dave Dice, Yossi Lev, Victor Luchangco, Virendra
  J. Marathe, and Nir Shavit.  2013.  NUMA-aware reader-writer locks.
  SIGPLAN Not. 48, 8 (August 2013), 157–166.
  DOI:https://doi.org/10.1145/2517327.24425

(The full article available from:
  http://mcg.cs.tau.ac.il/papers/ppopp2013-rwlocks.pdf)

The implementation is based on the The Writer-Preference Lock (C-RW-WP)
variant (see the 3.4 section of the paper for the rationale).  The
implemented algorithm has been modified for simplicity and for usage
patterns in rbtdb.c.  Namely, we haven't implemented the cohort locks
because that would require a knowledge of NUMA nodes, instead a simple
atomic_bool is used as synchronization point for writer lock.  Apart
from that the implementation uses writer_barrier that will raise after a
while when readers lock can't be acquired and separate ingress and
egress readers counters queues to reduce both inter and intra-thread
contention.

This (C-)RW-WP implementation is available by calling
isc_rwlock_init_ex() with ISC_RWLOCK_IMPL_RW_WP passed as chosen
implementation.  The isc_rwlock_init() will then use system
pthread_rwlocks in places where the additional complexity is not needed.
As the (C-)RW-WP uses cache-aligned ingress and egress readers counters,
it can be quite memory hungry, which makes it unsuitable for general
use.  As such it's only used in the rbtdb.c and all other rwlocks are
served by using pthread rwlock.  All the usage scenarios have been
reviewed and the only place where we use isc_rwlock_tryupgrade() and
isc_rwlock_downgrade() and which is also sensitive to writers starvation
is indeed rbtdb.c.