Do not attempt to resign the "dynamic" zone
In the "masterformat" system test, ns3/dynamic.db.map is a zone file in
"map" format that contains a zone called "dynamic".  That zone file
contains RRSIG records with expiration times dating back to 2005.  Since
the "dnssec-update-mode" option defaults to "maintain", named attempts
to refresh those expired DNSSEC signatures when the "dynamic" zone is
loaded after startup.  However, no signing keys exist for the "dynamic"
zone.  Since not finding any signing keys is not considered an error (as
keys may have been e.g. intentionally put offline), zone_resigninc()
calls set_resigntime() as if it hit a resigning quantum.  This results
in zone_resigninc() being immediately rescheduled and called again as
soon as possible, effectively causing a tight loop which is not broken
until signing keys for the zone are found.  As testing zone resigning is
not the purpose of the "masterformat" system test, prevent this problem
from getting triggered by setting "dnssec-update-mode" for the "dynamic"
zone to "no-resign".