schedule queues to run as uv_idle_t events
Currently, the scheduling via the queue might lead to an endless loop in
case that the netievent will schedule another event to be run on the
same loop that we've been processing in the async handler.

We don't have ability to quickly move all scheduled netievents to the
duplicate queue thus instead of processing the netivents in the async_cb
directly, we just reschedule them to be run as the uv_idle_t on the
worker uv_loop_t.