Cleanup function calls from assertion macros
When we changed the DbC assertion macros to simple assert() for
CPPCHECK, couple of assertWithSideEffect defects popped up.  Those are
all false positive because we don't use assert() in production code, but
it's easy to clean this up, so this commit moves all the function calls
that cppcheck identified inside the assertion macros outside.