Silence Gcc Analyzer false positives in mem.c
In mem_allocate(), we return "middle" of the allocated memory and keep
the first sizeinfo chunks for keeping the size of the allocated chunk
and for debugging.  This confuses Gcc Analyzer, so we just disable the
warning for the mem_allocate() function.

In mem_create(), the default_memallocate() actually can't return NULL
because the size of the isc_mem_t is not ever going to be 0, but the
extra INSIST() doesn't cause any penalty, so let's just add it and be
over with it.