create a clientmgr resource pool, use it for client messages
we want a single resource pool per clientmgr, one for each worker
thread. this should speed up the creation of new message objects.
note that this change by itself shouldn't have much affect on
performance, since client messages are reused until the server is
shut down, and only need to be created when the server is first
starting up. later, we'll be pass the resource pools in other
situations where messages are created (e.g., the resolver), and
that should help more.