implement serve-stale with netmgr timeouts
when we're setting the retry timeout for a recursive query, we check
to see if the serve-stale timeout should occur before the retry timer;
if so, we use that for the query timeout instead.

the resquery_timeout() function checks to see whether it's time to
serve stale data, and if so it does so, and then resumes waiting for
what would have been the retry timeout otherwise.