Reduce the thread scheduling priority during zone dumping
After the move of the zone dumping process, we can slightly reduce the
scheduling priority of the zone dumping process.  This is either done by
setting the SCHED_BATCH priority at the beginning and SCHED_OTHER (aka
SCHED_NORMAL) back when the zone dumping is done.

(On Windows, this is done by using THREAD_MODE_BACKGROUND_BEGIN and
THREAD_MODE_BACKGROUND_END combo.)

The description (below, from the man page) of SCHED_BATCH talks about
CPU-intensive work which is not the case of the only usage so far, but
we can't use SCHED_IDLE, and the extra penalty in scheduling is what we
need here.

     SCHED_BATCH: Scheduling batch processes

       SCHED_BATCH can be used only at static priority 0.  This policy
       is similar to SCHED_OTHER in that it schedules the thread
       according to its dynamic priority (based on the nice value).  The
       difference is that this policy will cause the scheduler to always
       assume that the thread is CPU-intensive.  Consequently, the
       scheduler will apply a small scheduling penalty with respect to
       wakeup behavior, so that this thread is mildly disfavored in
       scheduling decisions.

       This policy is useful for workloads that are noninteractive, but
       do not want to lower their nice value, and for workloads that
       want a determin‐ istic scheduling policy without interactivity
       causing extra preemptions (between the workload's tasks).

The reason why we can't use SCHED_IDLE is twofold:

1. The rules when we can actually reset the scheduling priority back to
   SCHED_OTHER are complicated:

    * Special rules apply for the SCHED_IDLE policy.  In Linux kernels
      before 2.6.39, an unprivileged thread operating under this policy
      cannot change its policy, regardless of the value of its
      RLIMIT_RTPRIO resource limit.  In Linux kernels since 2.6.39, an
      unprivileged thread can switch to either the SCHED_BATCH or the
      SCHED_OTHER policy so long as its nice value falls within the
      range permitted by its RLIMIT_NICE resource limit (see
      getrlimit(2)).

2. The zone dumping process would become so slow that it won't dump the
   zone in a reasonable amount of time.