fix cancellation of dispatch entries
dispentries are now attached before connecting and detached
afterward to prevent a condition in which dns_dispatch_removeresponse()
was run after the connect callback event was already enqueued, causing
unpredictable behavior when the callback ran.

dns_dispatch_cancel() now flags a dispentry as having been canceled,
so that if the connect callback runs after cancellation, it will not
initiate a read.