Drop unused "today" Sphinx variable
The "today" Sphinx variable is defined in Makefile.docs and also used in
the sphinx-build invocation for building the ARM in the EPUB format.
However, that variable is not used in the ARM.  Remove the redundant
definitions of the "today" Sphinx variable.

Note that the "today" variable is still necessary for building the man
pages and therefore the definition of "today" in man_SPHINXOPTS is left
intact.