Drop unused "today" Sphinx variable
The "today" Sphinx variable is defined in doc/arm/Makefile.in.  However,
that variable is not used in the ARM.  Remove the redundant definition
of the "today" Sphinx variable.

Note that the "today" variable is still necessary for building the man
pages and therefore the definition of "today" in man_SPHINXOPTS is left
intact.

(cherry picked from commit ee65c0fb138cdf05f84aac07ff7e37248ae40f20)