Add log message when rehashing RBT hash table
Rehashing the hash table could have significant impact on the
performance of the server.  This commit adds a log message to indicate
the rehashing is in progress.

Add a 'name' variable to the RBT to be used for referencing the tree
its hash table. Most names are constant strings, but for rbtdb a bit
more effort is done to construct useful names. The RBT name will
include:
1. the rbt type (tree, nsec, nsec3)
2. whether the db represenets a zone or a cache
3. the zone origin