dig wasn't always running in background when it should
some of the serve-stale test cases are based groups of
dig calls running simultaneously in the background: the
test pauses and resumes running after 'wait'. in some cases
the final call to dig in a group wasn't in the background,
and this sometimes caused delays that affected later
test results.  in another case, a test was simplified
and made more reliable by running dig in the foreground
removing a sleep.

the extension of the dig timeout period from 10 to
11 seconds in commit 5307bf64ce80 was left undone in a
few places and has now been completed.